#333 docs: stop.sh 스크립트 수정

+++ scripts/stop.sh
@@ -2,9 +2,17 @@
 # scripts/stop.sh
 
 if [ "$DEPLOYMENT_GROUP_NAME" == "release" ]; then
-    cd /home/ec2-user/build/release || exit
-    sudo docker-compose -f /home/ec2-user/build/release/docker-compose-release.yml down -v
+    if [ -f /home/ec2-user/build/release/docker-compose-release.yml ]; then
+        cd /home/ec2-user/build/release || exit
+        sudo docker-compose -f /home/ec2-user/build/release/docker-compose-release.yml down -v
+    else
+        echo "docker-compose-release.yml 파일이 존재하지 않습니다."
+    fi
 elif [ "$DEPLOYMENT_GROUP_NAME" == "dev" ]; then
-    cd /home/ec2-user/build/dev || exit
-    sudo docker-compose -f /home/ec2-user/build/dev/docker-compose-dev.yml down -v
+    if [ -f /home/ec2-user/build/dev/docker-compose-dev.yml ]; then
+        cd /home/ec2-user/build/dev || exit
+        sudo docker-compose -f /home/ec2-user/build/dev/docker-compose-dev.yml down -v
+    else
+        echo "docker-compose-release.yml 파일이 존재하지 않습니다."
+    fi
 fi
